# add caching for cudnn descriptors