There are some steps for setting up some descriptors in the `Convolution` class. Furthermore, memory is allocated that is needed as "workspace" during a convolution. In the naive implementation, these steps were executed repeatedly for every application of a convolution. However, the settings change only when the convolution needs to be applied to images with different resolution, which barely happens when processing a video clip. 

Therefore, this PR introduces a caching mechanism that removes this redundancy.

Speedup roughly 2-3% on the test clip, which will be published on youtube in a later step.

Before:


Elapsed time in nanoseconds:
|                       | Total         | per frame
|-|-|-|
| incl. io              | 5260741121    | 20959127
| excl. io              | 2015754933    | 8030896
| gpu                   | 1509760936    | 6014983
| w/o conv. to int      | 1488816208    | 5931538

after:

|                        | Total        |  per frame
|-|-|-|
| incl. io               | 4944791394   |  19700364
| excl. io               | 1908645536   |  7604165
| gpu                    | 1460924323   |  5820415
| w/o conv. to int       | 1448427256   |  5770626